(BIDS-2859) useRefBdridge shorter and more versatile
